Make IDs retain data on 64-bit platforms.

+++ src/elkhound/c/c_type.h
@@ -46,7 +46,7 @@ public:     // funcs
   virtual ~AtomicType();
 
   // stand-in if I'm not really using ids..
-  long getId() const { return (long)this; }
+  uintptr_t getId() const { return (uintptr_t)this; }
 
   virtual Tag getTag() const = 0;
   bool isSimpleType() const { return getTag() == T_SIMPLE; }
